+++ recpack/tests/test_data/test_datasets.py
@@ -163,7 +163,7 @@ def test_recsys_challenge_2015(path):
 
     data = d.load_interaction_matrix()
 
-    assert data.shape == (54, 291)
+    assert data.shape == (83, 26)
 
     # We can't just download this dataset, since it requires Kaggle access
     with pytest.raises(NotImplementedError):